LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-key-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=date id=date value="2000-01-02">
<script>
description("Tests if calendar picker key bindings work as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('date'), test1);

function test1() {
    debug('Check that page popup exists.');
    shouldBe('popupWindow.pagePopupController.toString()', '"[object PagePopupController]"');

    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-02');

    debug('Check that arrow keys work properly even when going between weeks and months.');

    debug('press up');
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug('press up');
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-19');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-19');
    
    debug('press down');
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug('press down');
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-02');

    debug('press up');
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    // Move from top left of January 2000 to bottom right of December 1999.
    debug('press left');
    eventSender.keyDown('leftArrow');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-25');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-25');

    debug('press right');
    eventSender.keyDown('rightArrow');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press m");
    eventSender.keyDown('m');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-23');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-23');

    debug("press shift + m");
    eventSender.keyDown('M');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press y");
    eventSender.keyDown('y');
    shouldBeEqualToString('currentMonth()', '2000-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2000-12-24');
    shouldBeEqualToString('highlightedDayCells()', '2000-12-24');

    debug("press shift + y");
    eventSender.keyDown('Y');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press d");
    eventSender.keyDown('d');
    shouldBeEqualToString('currentMonth()', '2009-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2009-12-27');
    shouldBeEqualToString('highlightedDayCells()', '2009-12-27');

    debug("press shift + d");
    eventSender.keyDown('D');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press enter");
    eventSender.keyDown('\n');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '1999-12-26');
    shouldBeEqualToString('selectedDayCells()', '1999-12-26');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');
    shouldBeEqualToString('document.getElementById("date").value', '1999-12-26');

    // Test twice in case midnight.
    shouldBeTrue('testToday() || testToday()');

    debug('Check if tabbing works.');
    shouldBeEqualToString('focusedElement()', '.list-view calendar-table-view');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.calendar-navigation-button');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.calendar-navigation-button today-button');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.calendar-navigation-button');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.month-popup-button');

    debug("press esc");
    eventSender.keyDown('\x1B');
    waitUntilClosing(test2AfterClosing);
}

function test2AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');

    finishJSTest();
}

function testToday() {
    eventSender.keyDown('t');
    var now = new Date();
    var expectedDate = serializeDate(now.getFullYear(), now.getMonth(), now.getDate());
    var expectedMonth = (new popupWindow.Month(now.getFullYear(), now.getMonth())).toString();
    return selectedValue() === expectedDate &&
           highlightedValue() === expectedDate &&
           currentMonth() === expectedMonth &&
           document.getElementById("date").value === expectedDate;
}

function serializeDate(year, month, day) {
    var yearString = String(year);
    if (yearString.length < 4)
        yearString = ("000" + yearString).substr(-4, 4);
    return yearString + "-" + ("0" + (month + 1)).substr(-2, 2) + "-" + ("0" + day).substr(-2, 2);
}

function focusedElement() {
    var element = popupWindow.document.activeElement;
    if (!element)
        return null;
    var identifier = "." + element.className;
    if (element.value)
        identifier += "[value=" + element.value + "]";
    return identifier;
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
